requirements.txt
py-desmume>=0.0.9
anthropic>=0.84.0
pydantic>=2.0
Pillow>=10.0
numpy>=1.24
